version 2.10 (user achievement badge)

+++ database/migrations/2017_04_23_201812_create_posts_table.php
@@ -17,9 +17,9 @@ class CreatePostsTable extends Migration
             $table->increments('id');
             $table->boolean('status')->default(1);
             $table->integer('user_id')->unsigned();
-            $table->integer('views')->unsigned()->default(0);
-            $table->integer('likes')->unsigned()->default(0);
-            $table->integer('bookmarks')->unsigned()->default(0);
+            $table->bigInteger('views')->unsigned()->default(0);
+            $table->bigInteger('likes')->unsigned()->default(0);
+            $table->bigInteger('bookmarks')->unsigned()->default(0);
             $table->string('title');
             $table->string('featured_image')->nullable();
             $table->string('image');
